Reorganize `util` outside components dir
