+++ package.json
@@ -9,6 +9,7 @@
     "axios": "^0.16.2",
     "prop-types": "^15.5.10",
     "react": "^15.6.1",
+    "react-center": "^1.1.1",
     "react-dom": "^15.6.1",
     "react-loading": "^0.1.4",
     "react-router-dom": "^4.2.2"
