+++ src/lib/components/ProjectDetail.js
@@ -1,14 +1,16 @@
-import React from 'react'
 import PropTypes from 'prop-types'
 
 // Helper to find a project by a slug name
-import renderProjectBySlug from './util/helpers/projectBySlug'
+import renderProjectBySlug from '../util/helpers/projectBySlug'
 
 // Helper to determine if a slug was provided on route params
-import slugProvided from './util/helpers/slugProvided'
+import slugProvided from '../util/helpers/slugProvided'
 
-// Use the shared error handlers
-import {error, loading} from './util/errors'
+// Use the shared error handler
+import {error} from '../util/handlers/errors'
+
+// Use the shared loading handler
+import {loading} from '../util/handlers/loaders'
 
 // Handles the showing of the ProjectDetail
 function ProjectDetail (props) {
