+++ src/lib/util/handlers/errors/index.js
@@ -2,4 +2,3 @@ import React from 'react'
 
 export const notFound = (slug) => (<p>Project not found: {slug}</p>)
 export const error    = ()     => (<p>Error</p>)
-export const loading  = ()     => (<p>Loading</p>)
